requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7